% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_g4s_assessment_functions.R
\name{gfs_assessment_markbooks}
\alias{gfs_assessment_markbooks}
\title{Get markbooks.}
\usage{
gfs_assessment_markbooks(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the structure of markbooks in the chosen academic year, in terms of
the marksheets they contain, and the markslots within each marksheet.
}
\examples{
gfs_assessment_markbooks(2020)
}
