% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_g4s_assessment_functions.R
\name{gfs_assessment_marks}
\alias{gfs_assessment_marks}
\title{Get marks.}
\usage{
gfs_assessment_marks(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of the marks achieved by students for individual marks.
\itemize{
\item Use \code{gfs_assessment_markbooks()} to resolve markbooks.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_assessment_marks(2020)
}
