% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_g4s_assessment_functions.R
\name{gfs_assessment_marksheets}
\alias{gfs_assessment_marksheets}
\title{Get marksheets.}
\usage{
gfs_assessment_marksheets(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of the grades achieved by students for individual marksheets.
\itemize{
\item Use \code{gfs_assessment_markbooks()} to resolve markbooks.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_assessment_marksheet_grades(2020)
}
