% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_g4s_attainment_functions.R
\name{gfs_attainment_exam_results}
\alias{gfs_attainment_exam_results}
\title{Get exam results.}
\usage{
gfs_attainment_exam_results(academicYear, yearGroup)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}
}
\description{
Returns details of the exam results achieved by students.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
\item Use \code{gfs_teaching_subjects()} to resolve subject details.
}
}
\examples{
gfs_attainment_exam_results(2020, "11")
}
