% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_g4s_attainment_functions.R
\name{gfs_attainment_grade_types}
\alias{gfs_attainment_grade_types}
\title{Get grade types.}
\usage{
gfs_attainment_grade_types(academicYear, yearGroup)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}
}
\description{
Returns details of the available types of headline grades.
}
\examples{
gfs_attainment_grade_types(2020, "11")
}
