% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_g4s_attainment_functions.R
\name{gfs_attainment_grades}
\alias{gfs_attainment_grades}
\title{Get grades.}
\usage{
gfs_attainment_grades(academicYear, yearGroup)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}
}
\description{
Returns details of the headline grades for students.
\itemize{
\item Use \code{gfs_attainment_grade_types()} to resolve grade names.
\item Use \code{gfs_student_details()} to resolve student personal details.
\item Use \code{gfs_teaching_subjects()} to resolve subject details.
}
}
\examples{
gfs_attainment_grades(2020, "11")
}
