% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_g4s_attainment_functions.R
\name{gfs_attainment_prior}
\alias{gfs_attainment_prior}
\title{Get prior attainment.}
\usage{
gfs_attainment_prior(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns prior attainment for students.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_attainment_prior(2020)
}
