% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_g4s_attendance_functions.R
\name{gfs_attendance_codes}
\alias{gfs_attendance_codes}
\title{Get attendance codes.}
\usage{
gfs_attendance_codes(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the Attendance codes and any aliases being used for the
academic year.
}
\examples{
gfs_attendance_codes(2020)
}
