% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_g4s_attendance_functions.R
\name{gfs_attendance_student_summary}
\alias{gfs_attendance_student_summary}
\title{Get student attendance summary.}
\usage{
gfs_attendance_student_summary(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns summary session data for all students for the academic
year up to (but not including) today.
\itemize{
\item Use \code{gfs_attendance_codes} to resolve attendance codes.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_attendance_student_summary(2020)
}
