% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_g4s_behaviour_functions.R
\name{gfs_behaviour_classification}
\alias{gfs_behaviour_classification}
\title{Get classifications.}
\usage{
gfs_behaviour_classification(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the current behaviour classification groups for the
supplied academic year.
}
\examples{
gfs_behaviour_classification(2020)
}
