% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_g4s_behaviour_functions.R
\name{gfs_behaviour_event_types}
\alias{gfs_behaviour_event_types}
\title{Get event types.}
\usage{
gfs_behaviour_event_types(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the current definition of the behaviour types for the
supplied academic year.
}
\examples{
gfs_behaviour_event_types(2020)
}
