% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_g4s_behaviour_functions.R
\name{gfs_behaviour_events_range}
\alias{gfs_behaviour_events_range}
\title{Get events over a range of dates.}
\usage{
gfs_behaviour_events_range(academicYear, goDateStart, goDateEnd)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDateStart}{start date of range as string in the form yyyy-mm-dd.}

\item{goDateEnd}{end date of range as string in the form yyyy-mm-dd.}
}
\description{
Returns list of events over a range of dates within
an academic year.
\itemize{
\item Use \code{gfs_behaviour_event_types()} to resolve event characteristics.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_behaviour_events_range(2020, "2020-03-16", "2020-03-19")
}
