% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_g4s_behaviour_functions.R
\name{gfs_behaviour_followup}
\alias{gfs_behaviour_followup}
\title{Get follow-up actions.}
\usage{
gfs_behaviour_followup(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the current definitions for follow up behaviour actions for the
supplied academic year.
}
\examples{
gfs_behaviour_followup(2020)
}
