% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_g4s_behaviour_functions.R
\name{gfs_behaviour_immediate}
\alias{gfs_behaviour_immediate}
\title{Get immediate actions.}
\usage{
gfs_behaviour_immediate(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the current definitions for immediate behaviour actions for the
supplied academic year.
}
\examples{
gfs_behaviour_immediate(2020)
}
