% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_g4s_timetables_functions.R
\name{gfs_calendar}
\alias{gfs_calendar}
\title{Get calendar.}
\usage{
gfs_calendar(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of when the school is open, closed or open for staff
training only in the chosen academic year.
\itemize{
\item Use \code{gfs_timetables()} to resolve date, week, and period details.
}
}
\examples{
gfs_calendar(2020)
}
