% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_g4s_timetables_functions.R
\name{gfs_classes}
\alias{gfs_classes}
\title{Get classes.}
\usage{
gfs_classes(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of all timetabled classes in the chosen academic year.
\itemize{
\item Use \code{gfs_timetables()} to resolve date, week, and period details.
\item Use \code{gfs_teaching_teachers()} to resolve teacher details.
}
}
\examples{
gfs_classes(2020)
}
