% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_g4s_clean_attainment_functions.R
\name{gfs_clean_attainment}
\alias{gfs_clean_attainment}
\title{Get clean attainment data.}
\usage{
gfs_clean_attainment(academicYear, yearGroup)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}
}
\description{
Returns clean details of attainment in the chosen academic year for chosen
year group.
}
\examples{
gfs_clean_attainment(2020, "10")
}
