% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_g4s_clean_attainment_functions.R
\name{gfs_clean_attainment_multiple}
\alias{gfs_clean_attainment_multiple}
\title{Get clean attainment data for multiple year groups.}
\usage{
gfs_clean_attainment_multiple(academicYear, yearGroupFrom, yearGroupTo)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroupFrom}{starting year group as string.}

\item{yearGroupTo}{ending year group as string.}
}
\description{
Returns clean details of attainment in the chosen academic year for multiple
year groups.
}
\examples{
gfs_clean_attainment_multiple(2021, "7", "11")
}
