% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_g4s_clean_attendance_functions.R
\name{gfs_clean_attendance_student_session}
\alias{gfs_clean_attendance_student_session}
\title{Get clean student session marks.}
\usage{
gfs_clean_attendance_student_session(academicYear, goDate, session = NULL)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDate}{date as string in the form yyyy-mm-dd.}

\item{session}{session as string.}
}
\description{
Returns clean details of student session marks for chosen
academic year, date, and session.
}
\examples{
gfs_clean_attendance_student_session(2020, "2020-07-01")
gfs_clean_attendance_student_session(2020, "2020-07-01", "am")
}
