% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_g4s_clean_attendance_functions.R
\name{gfs_clean_attendance_student_session_range}
\alias{gfs_clean_attendance_student_session_range}
\title{Get clean student session marks over a range of dates.}
\usage{
gfs_clean_attendance_student_session_range(
  academicYear,
  goDateStart,
  goDateEnd,
  session = NULL
)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{goDateStart}{start date of range as string in the form yyyy-mm-dd.}

\item{goDateEnd}{end date of range as string in the form yyyy-mm-dd.}

\item{session}{session as string.}
}
\description{
Returns clean student session marks over a range of dates for chosen
academic year.
}
\examples{
gfs_clean_attendance_student_session_range(2021, "2021-07-12", "2021-07-16")
gfs_clean_attendance_student_session_range(2021, "2021-07-12", "2021-07-16", session = "am")
}
