% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_g4s_clean_attendance_functions.R
\name{gfs_clean_attendance_student_summary}
\alias{gfs_clean_attendance_student_summary}
\title{Get clean student attendance summary.}
\usage{
gfs_clean_attendance_student_summary(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean details of student attendance summary for chosen
academic year.
}
\examples{
gfs_clean_attendance_student_summary(2020)
}
