% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_g4s_clean_behaviour_functions.R
\name{gfs_clean_behaviour_events_range}
\alias{gfs_clean_behaviour_events_range}
\title{Get clean behaviour events.}
\usage{
gfs_clean_behaviour_events_range(academicYear, goDateStart, goDateEnd)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}

\item{type}{qualification type as string.}
}
\description{
Returns clean behaviour events over a range of dates within in the chosen
academic year.
}
\examples{
gfs_clean_behaviour_events_range(2021, "2021-07-12", "2021-07-16")
}
