% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_g4s_clean_class_list_functions.R
\name{gfs_clean_class_list_student}
\alias{gfs_clean_class_list_student}
\title{Get student-focused class list.}
\usage{
gfs_clean_class_list_student(academicYear, student)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{student}{student search string.}
}
\description{
Returns clean details of all classes in the chosen academic year for chosen
student.
}
\examples{
gfs_class_list_student(2020, "smith")
}
