% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_g4s_clean_class_list_functions.R
\name{gfs_clean_class_list_teacher}
\alias{gfs_clean_class_list_teacher}
\title{Get teacher-focused class list.}
\usage{
gfs_clean_class_list_teacher(academicYear, staffCode = NULL)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{staffCode}{case-sensitive staff code as string.}
}
\description{
Returns clean details of all classes in the chosen academic year for chosen
staff.  If staff code is not provided, class list for all staff are returned.
}
\examples{
gfs_clean_class_list_teacher(2020, "ACH")
gfs_clean_class_list_teacher(2020)
}
