% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_g4s_clean_attainment_functions.R
\name{gfs_clean_exam_results}
\alias{gfs_clean_exam_results}
\title{Get external examination results.}
\usage{
gfs_clean_exam_results(academicYear, yearGroup, type = NULL)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{yearGroup}{year group as string.}

\item{type}{qualification type as string.}
}
\description{
Returns clean details of external examination results in the chosen
academic year for chosen
year group.
}
\examples{
gfs_clean_exam_results(2020, "11")
gfs_clean_exam_results(2020, "11", type = "gcse")
gfs_clean_exam_results(2020, "11", type = c("gcse", "btec"))
}
