% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_g4s_clean_school_population_functions.R
\name{gfs_clean_school_population}
\alias{gfs_clean_school_population}
\title{Get clean school population data.}
\usage{
gfs_clean_school_population()
}
\description{
Returns clean details of school population for all available academic years,
along with a line plot.
}
\examples{
gfs_clean_school_population()
}
