% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_g4s_clean_student_details_functions.R
\name{gfs_clean_student_details}
\alias{gfs_clean_student_details}
\title{Get clean student details.}
\usage{
gfs_clean_student_details(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns clean student details for chosen academic year.
}
\examples{
gfs_clean_student_details(2021)
}
