% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_g4s_clean_student_details_functions.R
\name{gfs_clean_student_send_search}
\alias{gfs_clean_student_send_search}
\title{Get clean SEND notes search.}
\usage{
gfs_clean_student_send_search(academicYear, notesSearch)
}
\arguments{
\item{academicYear}{notesSearch as string.}
}
\description{
Returns clean student details based on SEND notes search for chosen
academic year.
}
\examples{
gfs_clean_student_send_search(2021, "extra time")
gfs_clean_student_send_search(2021, "irlen")
}
