% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_g4s_report_functions.R
\name{gfs_reports}
\alias{gfs_reports}
\title{Get reports.}
\usage{
gfs_reports(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of available progress reports in the chosen academic year,
including the grades and attributes associated with each set of reports.
}
\examples{
gfs_reports(2021)
}
