% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_g4s_report_functions.R
\name{gfs_reports_grades}
\alias{gfs_reports_grades}
\title{Get report grades}
\usage{
gfs_reports_grades(academicYear, reportID)
}
\arguments{
\item{academicYear}{academic year as integer.}

\item{reportID}{report ID as integer.}
}
\description{
Returns grades for students in chosen academic year and report.
\itemize{
\item Use \code{gfs_reports} to resolve name and points value for an attribute choice
for the student.
\item Use \code{gfs_teaching_subjects()} to resolve subject details.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_reports_grades(2021, 57102)
}
