% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_g4s_setup.R
\name{gfs_setup}
\alias{gfs_setup}
\title{Run set up for gfs functions.}
\usage{
gfs_setup(api_key)
}
\arguments{
\item{api_key}{API key as string.}
}
\description{
Use this function to store your API key for use with gfs functions.
Authentication is required for all requests.  See your school's
administrator for an API key.
}
\examples{
gfs_setup("a96650481f0245eea396726f85ac7049")
}
