% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_g4s_timetables_functions.R
\name{gfs_student_classes}
\alias{gfs_student_classes}
\title{Get student classes.}
\usage{
gfs_student_classes(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of which students are linked to which timetabled classes in
the chosen academic year.
\itemize{
\item Use \code{gfs_classes()} to resolve class details.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_classes(2020)
}
