% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_demographics}
\alias{gfs_student_demographics}
\title{Get demographic attributes.}
\usage{
gfs_student_demographics(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the following attributes for students for the chosen academic year:
country of birth, home language, first language, and nationality.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_demographics(2020)
}
