% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_details}
\alias{gfs_student_details}
\title{Get students details.}
\usage{
gfs_student_details(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the current personal details (DOB, gender, names) of the students
that were at school in the chosen academic year.
}
\examples{
gfs_student_details(2020)
}
