% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_edu_details}
\alias{gfs_student_edu_details}
\title{Get education details.}
\usage{
gfs_student_edu_details(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns education details for students.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_edu_details(2020)
}
