% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_medical}
\alias{gfs_student_medical}
\title{Get student medical conditions.}
\usage{
gfs_student_medical(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns medical conditions for students.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_medical(2020)
}
