% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_photos}
\alias{gfs_student_photos}
\title{Get student photos.}
\usage{
gfs_student_photos(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns photos for students.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_photos(2020)
}
