% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_g4s_student_functions.R
\name{gfs_student_send}
\alias{gfs_student_send}
\title{Get SEND attributes.}
\usage{
gfs_student_send(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns the following attributes for students for the chosen academic year:
SEN code, SEN status, educational needs, gifted/talented, gifted/talented
subject.
\itemize{
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_student_send(2020)
}
