% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_departments}
\alias{gfs_teaching_departments}
\title{Get departments.}
\usage{
gfs_teaching_departments(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of departments in the chosen academic year.
}
\examples{
gfs_teaching_departments(2020)
}
