% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_groups}
\alias{gfs_teaching_groups}
\title{Get groups.}
\usage{
gfs_teaching_groups(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of teaching groups for each subject in the chosen academic
year.
\itemize{
\item Use \code{gfs_teaching_subjects()} to resolve subject details.
}
}
\examples{
gfs_teaching_groups(2020)
}
