% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_groups_students}
\alias{gfs_teaching_groups_students}
\title{Get groups and students.}
\usage{
gfs_teaching_groups_students(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of the students in each teaching group in the chosen
academic year.
\itemize{
\item Use \code{gfs_teaching_groups()} to resolve group details.
\item Use \code{gfs_student_details()} to resolve student personal details.
}
}
\examples{
gfs_teaching_groups_students(2020)
}
