% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_groups_teachers}
\alias{gfs_teaching_groups_teachers}
\title{Get groups and teachers.}
\usage{
gfs_teaching_groups_teachers(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of teachers assigned to each teaching group in the chosen
academic year.
\itemize{
\item Use \code{gfs_teaching_groups()} to resolve group details.
\item Use \code{gfs_teaching_teachers()} to resolve teacher details.
}
}
\examples{
gfs_teaching_groups_teachers(2020)
}
