% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_subjects}
\alias{gfs_teaching_subjects}
\title{Get subjects.}
\usage{
gfs_teaching_subjects(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns all subjects in the chosen academic year.
\itemize{
\item Use \code{gfs_teaching_departments()} to resolve department details.
}
}
\examples{
gfs_teaching_subjects(2020)
}
