% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_g4s_teaching_functions.R
\name{gfs_teaching_teachers}
\alias{gfs_teaching_teachers}
\title{Get teachers.}
\usage{
gfs_teaching_teachers(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of the teachers in the chosen academic year.
}
\examples{
gfs_teaching_teachers(2020)
}
