% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_g4s_timetables_functions.R
\name{gfs_timetables}
\alias{gfs_timetables}
\title{Get timetables.}
\usage{
gfs_timetables(academicYear)
}
\arguments{
\item{academicYear}{academic year as integer.}
}
\description{
Returns details of the timetables and their associated periods for the
chosen academic year.
}
\examples{
gfs_timetables(2020)
}
